requirements_v1.txt
torch
torchvision
opencv-python
numpy
Pillow
DateTime
matplotlib==3.8.0
plyfile
HTML4Vision
tensorboardX
imgaug
iopath
imagecorruptions
timm
termcolor